+++ frontend/app/api/retrieve/meeting/[id]/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest } from "next/server";
 import { getAuth } from "../../../../../services/auth";
 import { prisma } from "../../../../../lib/prisma";
-import { toPublicMeeting } from "../../../../../util/publicMeeting";
+import { toPublicMeeting } from "../../../../../util/meetings/publicMeeting";
 import { getUnresolvedSuspension } from "../../../../../util/suspension";
 import { formatETDateString } from "../../../../../util/date/timeUtils";
 const getMeeting = async(request: NextRequest) => {
