+++ frontend/app/components/admin/diagnostics/ConflictList.tsx
@@ -9,11 +9,11 @@ import {
   ConflictListRow,
   formatOverlapSummary,
   formatMeetingSchedule,
-} from "../../../../util/conflictDisplay";
+} from "../../../../util/meetings/conflictDisplay";
 import styles from "../../../../styles/components/admin/ConflictList.module.scss";
 
-export type { ConflictRecurrenceSummary, ConflictMeetingSummary, ConflictListRow } from "../../../../util/conflictDisplay";
-export { fieldLabel, formatOverlapSummary, formatMeetingSchedule } from "../../../../util/conflictDisplay";
+export type { ConflictRecurrenceSummary, ConflictMeetingSummary, ConflictListRow } from "../../../../util/meetings/conflictDisplay";
+export { fieldLabel, formatOverlapSummary, formatMeetingSchedule } from "../../../../util/meetings/conflictDisplay";
 
 interface ConflictListProps {
   conflicts: ConflictListRow[];
