+++ frontend/app/components/calendar/desktop/WeekView.tsx
@@ -4,7 +4,7 @@ import DayColumn from "../shared/DayColumn";
 import { filterMeetingsForDate, MeetingFilters } from "../../../../util/filters/meetingFilters";
 import { ROOM_COLORS, ZOOM_ROOM_COLOR, REMOTE_COLOR } from "../../../../util/rooms/filterColors";
 import { formatETDateString } from "../../../../util/date/timeUtils";
-import { layoutOverlappingMeetings } from "../../../../util/meetingOverlapLayout";
+import { layoutOverlappingMeetings } from "../../../../util/meetings/meetingOverlapLayout";
 import { getFirstDayOfWeek, getDaysOfWeek } from "../../../../util/date/weekDates";
 import { useWeekMeetings, WeekMeeting } from "../../../../hooks/useWeekMeetings";
 import TopLoadingBar from "../../atoms/TopLoadingBar";
