+++ frontend/app/components/meeting-form/NewMeeting.tsx
@@ -17,7 +17,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { physicalRoomOptions, zoomRoomOptions } from "../../../util/rooms/rooms";
 import { useMeetingForm, CAL_TYPE_OPTIONS, CAL_TYPE_COLOR, DESCRIPTION_MAX_LENGTH } from '../../../hooks/useMeetingForm';
 import { IMeeting } from '../../../types/models';
-import { ConflictListRow } from '../../../util/conflictDisplay';
+import { ConflictListRow } from '../../../util/meetings/conflictDisplay';
 
 import styles from '../../../styles/components/meeting-form/MeetingForm.module.scss';
