+++ frontend/app/components/meeting-form/ViewMeeting.tsx
@@ -13,9 +13,9 @@ import { IRecurrencePattern } from '../../../types/models';
 import { formatCompactTimeRange, formatMeetingDateLine } from "../../../util/date/timeFormat";
 import { formatETDateString } from "../../../util/date/timeUtils";
 import { retryMeetingSync } from "../../../services/syncMeeting";
-import { formatSuspensionStatusText } from "../../../util/suspensionText";
+import { formatSuspensionStatusText } from "../../../util/meetings/suspensionText";
 import { ROOM_COLORS, ZOOM_ROOM_COLOR } from "../../../util/rooms/filterColors";
-import { formatRecurrencePattern } from "../../../util/recurrenceDisplay";
+import { formatRecurrencePattern } from "../../../util/meetings/recurrenceDisplay";
 import { isZoomRoomMismatched } from "../../../util/rooms/rooms";
 import { linkify } from "../../../util/common/linkify";
 import { zoomHostLabel } from "../../../util/rooms/zoomHosts";
