+++ frontend/hooks/useWeekMeetings.ts
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useRef, useState } from "react";
 import { formatETDateString } from "../util/date/timeUtils";
 import { createCache } from "../util/common/simpleCache";
 import { IMeeting } from "../types/models";
-import { OverlapMeeting } from "../util/meetingOverlapLayout";
+import { OverlapMeeting } from "../util/meetings/meetingOverlapLayout";
 
 export interface WeekMeeting extends OverlapMeeting {
     syncError?: boolean;
